+++ natural_selection/__init__.py
@@ -1,4 +1,4 @@
-__version__ = '0.2.12'
+__version__ = '0.2.16'
 
 def get_random_string(length : int = 8, include_numeric=False) -> str:
     """