+++ tests/unittest/ga/test_island.py
@@ -19,6 +19,19 @@ class TestSimpleIsland(unittest.TestCase):
 
         self.life = Island({'x': 0.6, 'y' : 0.2}, crossover_function=crossover_two_n_point)
 
+    def test_write_report(self):
+        self.life.initialise(self.ind, population_size=5)
+        self.life.evolve(crossover_params={'n_points': 2})
+
+        self.life.write_report('test.csv')
+        self.life.write_report('test.json', output_json=True)
+
+        self.assertTrue(os.path.isfile('test.csv'))
+        self.assertTrue(os.path.isfile('test.json'))
+
+        os.remove('test.csv')
+        os.remove('test.json')
+
     def test_create(self):
         self.life.initialise(self.ind, population_size=5)
         self.assertEquals(len(self.life.population), 5)